optimizate saas api variable naming

+++ src/main/java/com/cobo/custody/api/client/impl/CoboWeb3ApiRestClientImpl.java
@@ -75,7 +75,7 @@ public class CoboWeb3ApiRestClientImpl implements CoboWeb3ApiRestClient {
     }
 
     @Override
-    public ApiResponse<Void> withdraw(String coin, String requestId, String fromAddr, String toAddr, int amount) {
+    public ApiResponse<Void> withdraw(String coin, String requestId, String fromAddr, String toAddr, long amount) {
         return executeSync(coboWeb3ApiService.web3Withdraw(coin, requestId, fromAddr, toAddr, amount));
     }
 
@@ -87,7 +87,7 @@ public class CoboWeb3ApiRestClientImpl implements CoboWeb3ApiRestClient {
     @Override
     public ApiResponse<Void> contract(String chainCode, String requestId, String walletAddr,
                                       String contractAddr, String methodId, String methodName,
-                                      String args, Integer amount) {
+                                      String args, Long amount) {
         return executeSync(coboWeb3ApiService.web3Contract(chainCode, requestId, walletAddr, contractAddr,
                 methodId, methodName, args, amount));
     }
